# API/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Blog Personal API
  version: 1.0.0
  description: API REST para un blog personal
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
paths:
...
  /api/posts:
    get:
      summary: Obtener posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Número de página
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Posts por página
        - in: query
          name: category
          schema:
            type: string
          description: Filtrar por categoría
      responses:
        200:
          description: Lista de posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
...
  /api/comments:
    post:
      summary: Crear comentario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - post_id
                - content
                - author_name
                - author_email
              properties:
                post_id:
                  type: integer
                content:
                  type: string
                author_name:
                  type: string
                author_email:
                  type: string
                  format: email
      responses:
        201:
          description: Comentario creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...
components:
  schemas:
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: string
        view_count:
          type: integer
        created_at:
          type: string
          format: date-time
        category_name:
          type: string
        author_name:
          type: string
...